refactor: extract chat session model
