+++ apps/desktop/src/plugins/builtin/ai_chat/chat_store.ts
@@ -34,6 +34,11 @@ import {
   serializeSessionForStorage,
 } from "./chat_session_persistence";
 import {
+  buildLegacyAcpHandoffPrompt,
+  runtimeMessagesFromChatMessages,
+  summarizeChatSessions,
+} from "./chat_session_model";
+import {
   chatSessionMatchesVaultRoot,
   filterForChatSessionVaultRoot,
   normalizeChatSessionVaultRoot,
@@ -71,18 +76,6 @@ const BUSY_SESSION_STATUSES: ChatSessionState["status"][] = [
   "awaiting-approval",
   "applying",
 ];
-
-type RuntimeChatMessage =
-  | { kind: "system"; content: string }
-  | { kind: "user"; content: string }
-  | { kind: "assistant"; content: string; toolCalls?: [] }
-  | {
-      kind: "toolResult";
-      callId: string;
-      toolName: string;
-      output: string;
-      isError: boolean;
-};
 
 let currentChatSessionVaultRoot: string | null = null;
 const pendingLegacyHandoffBySessionId = new Map<string, string>();
@@ -465,37 +458,6 @@ function appendSystemMessageOnce(sessionId: string, content: string): void {
   appendSystemMessage(sessionId, content);
 }
 
-function runtimeMessagesFromChatMessages(messages: ChatMessage[]): RuntimeChatMessage[] {
-  return messages.flatMap((message): RuntimeChatMessage[] => {
-    if (message.kind === "text") {
-      switch (message.role) {
-        case "system":
-          return [{ kind: "system", content: message.content }];
-        case "user":
-          return [{ kind: "user", content: message.content }];
-        case "assistant":
-          return [{ kind: "assistant", content: message.content, toolCalls: [] }];
-      }
-    }
-
-    if (message.kind === "tool") {
-      const output = message.output ?? message.error;
-      if (output == null) return [];
-      return [
-        {
-          kind: "toolResult",
-          callId: message.callId,
-          toolName: message.toolName,
-          output,
-          isError: message.success === false || message.error != null,
-        },
-      ];
-    }
-
-    return [];
-  });
-}
-
 function upsertAssistantPlaceholder(sessionId: string): string | null {
   const session = chatState.sessions[sessionId];
   if (!session) return null;
@@ -837,49 +799,15 @@ function resetToSession(
   persistChatSessions();
 }
 
-function sessionTitle(session: ChatSessionState): string {
-  const firstUserMessage = session.messages.find(
-    (message): message is ChatTextMessage => message.kind === "text" && message.role === "user",
-  );
-  const title = firstUserMessage?.content.trim();
-  if (title) {
-    return title.length > 64 ? `${title.slice(0, 61)}...` : title;
-  }
-
-  if (session.persistedTitle?.trim()) {
-    return session.persistedTitle.length > 64
-      ? `${session.persistedTitle.slice(0, 61)}...`
-      : session.persistedTitle;
-  }
-
-  switch (session.mode) {
-    case "agent":
-      return "Agent session";
-    case "inline":
-      return "Inline session";
-    case "ask":
-      return "Ask session";
-  }
-}
-
 function getSessionSummaries(): ChatSessionSummary[] {
   void chatState.sessionSummariesVersion;
-  const activeSessionId = chatState.activeSessionId;
   const vaultRoot = currentChatSessionVaultRoot;
-  return Object.values(chatState.sessions)
-    .filter((session) => normalizeChatSessionVaultRoot(session.workingDirectory) === vaultRoot)
-    .map((session) => ({
-      id: session.id,
-      agentId: session.agentId,
-      mode: session.mode,
-      title: sessionTitle(session),
-      draft: session.draft,
-      messageCount: session.messages.length,
-      status: session.status,
-      isActive: activeSessionId === session.id,
-      updatedAt: session.updatedAt,
-    }))
-    .sort((a, b) => b.updatedAt - a.updatedAt);
+  return summarizeChatSessions(
+    Object.values(chatState.sessions).filter(
+      (session) => normalizeChatSessionVaultRoot(session.workingDirectory) === vaultRoot,
+    ),
+    chatState.activeSessionId,
+  );
 }
 
 function switchSession(sessionId: string): boolean {
@@ -1013,32 +941,6 @@ function canRestoreSession(session: ChatSessionState): boolean {
     return true;
   }
   return typeof session.externalSessionId === "string" && session.externalSessionId.trim() !== "";
-}
-
-function buildLegacyAcpHandoffPrompt(session: ChatSessionState): string | null {
-  const transcript = session.messages
-    .flatMap((message) => {
-      if (message.kind !== "text") return [];
-      const content = message.content.trim();
-      if (!content) return [];
-      return `${message.role.toUpperCase()}: ${content}`;
-    })
-    .join("\n\n")
-    .trim();
-  if (!transcript) return null;
-
-  const maxLength = 12_000;
-  const clippedTranscript =
-    transcript.length > maxLength ? transcript.slice(transcript.length - maxLength) : transcript;
-
-  return [
-    "You are continuing a Kuku chat whose external ACP session could not be reattached.",
-    "Use the prior local transcript below as conversation context. Do not mention this handoff unless the user asks.",
-    "",
-    "<previous_transcript>",
-    clippedTranscript,
-    "</previous_transcript>",
-  ].join("\n");
 }
 
 async function restoreSession(sessionId: string): Promise<string | null> {
